Creazione prendi Utenti da finire

+++ src/main/java/it/unisa/diem/wordageddong17/model/Utente.java
@@ -24,7 +24,7 @@ public class Utente {
         this.tipo = tipo;
     }
 
-    public Utente(String username, String email, int punteggio, String giocatore) {
+    public Utente(String username, String email, float punteggio, String giocatore) {
         this.username = username;
         this.email = email;
         this.punteggioMigliore = punteggio;
